Moved all logical functions (any, allclose, all) into a new logical submodule
